Wow. Lot of code.
