--- includes/class-wc-coupon-restrictions-checkout-validation.php
@@ -1,1 +1,0 @@
-
