+++ includes/class-wc-coupon-restrictions-table.php
