+++ tests/phpunit/Integration/ApplyGenericCouponTest.php
@@ -5,8 +5,7 @@ use WP_UnitTestCase;
 use WC_Helper_Customer;
 use WC_Helper_Coupon;
 
-class Apply_Basic_Coupon_Test extends WP_UnitTestCase {
-
+class Apply_Generic_Coupon_Test extends WP_UnitTestCase {
 	/**
 	 * Tests that generic coupons can be applied.
 	 */
@@ -32,5 +31,4 @@ class Apply_Basic_Coupon_Test extends WP_UnitTestCase {
 		$coupon->delete();
 		$customer->delete();
 	}
-
 }
