+++ tests/phpunit/Integration/ApplyStateRestrictionCouponTest.php
@@ -5,8 +5,8 @@ use WP_UnitTestCase;
 use WC_Helper_Customer;
 use WC_Helper_Coupon;
 
-class Apply_Zipcode_Restriction_Test extends WP_UnitTestCase {
-
+class Apply_State_Restriction_Coupon_Test extends WP_UnitTestCase {
+	/** @var WC_Coupon */
 	public $coupon;
 	public $customer;
 	public $session;
@@ -14,8 +14,8 @@ class Apply_Zipcode_Restriction_Test extends WP_UnitTestCase {
 	public function setUp() {
 		// Creates a customer.
 		$customer = WC_Helper_Customer::create_customer();
-		$customer->set_billing_postcode( '78703' );
-		$customer->set_shipping_postcode( '78703' );
+		$customer->set_billing_state( 'TX' );
+		$customer->set_shipping_state( 'TX' );
 		$customer->save();
 		$this->customer = $customer;
 
@@ -45,13 +45,16 @@ class Apply_Zipcode_Restriction_Test extends WP_UnitTestCase {
 	}
 
 	/**
-	 * Tests applying a coupon with postcode restriction and valid customer.
+	 * Tests applying a coupon with country restriction and valid customer.
 	 */
-	public function test_postcode_restriction_with_valid_customer() {
+	public function test_coupon_state_restriction_with_valid_customer() {
 		$coupon = $this->coupon;
 
-		update_post_meta( $coupon->get_id(), 'postcode_restriction', '78703' );
+		// Apply state restriction to single state "TX"
+		update_post_meta( $coupon->get_id(), 'state_restriction', 'TX' );
 
+		// Adds a state restricted coupon.
+		// This should return true because customer billing is in TX.
 		$this->assertTrue( WC()->cart->apply_coupon( $coupon->get_code() ) );
 
 		// Verifies 1 coupon has been applied to cart.
@@ -59,33 +62,37 @@ class Apply_Zipcode_Restriction_Test extends WP_UnitTestCase {
 	}
 
 	/**
-	 * Tests applying a postcode restriction and non-valid customer.
+	 * Tests applying a coupon with a two state restriction and valid customer.
 	 */
-	public function test_coupon_country_restriction_with_nonvalid_customer() {
+	public function test_coupon_two_state_restriction_with_valid_customer() {
 		$coupon = $this->coupon;
 
-		update_post_meta( $coupon->get_id(), 'postcode_restriction', '000000' );
+		// Apply country restiction to two states "TX" and "CA"
+		update_post_meta( $coupon->get_id(), 'state_restriction', "ca,tx" );
 
+		// This should return true because customer billing is in tx.
+		$this->assertTrue( WC()->cart->apply_coupon( $coupon->get_code() ) );
+
+		// Verifies 1 coupon has been applied to cart.
+		$this->assertEquals( 1, count( WC()->cart->get_applied_coupons() ) );
+	}
+
+	/**
+	 * Tests applying a coupon with a two country restriction and non-valid customer.
+	 */
+	public function test_coupon_state_restriction_with_nonvalid_customer() {
+		$coupon = $this->coupon;
+
+		// Apply state restriction single state "CA".
+		update_post_meta( $coupon->get_id(), 'state_restriction', 'CA' );
+
+		// Adds a state restricted coupon.
+		// This should return false because customer billing is in TX.
 		$this->assertFalse( WC()->cart->apply_coupon( $coupon->get_code() ) );
 
 		// Verifies 0 coupons have been applied to cart.
 		$this->assertEquals( 0, count( WC()->cart->get_applied_coupons() ) );
 	}
-
-	/**
-	 * Tests applying a coupon with postcode restriction and valid customer.
-	 */
-	public function test_valid_postcode_restriction_wildcard() {
-		$coupon = $this->coupon;
-
-		update_post_meta( $coupon->get_id(), 'postcode_restriction', '00000,787*,ALPHAZIP' );
-
-		$this->assertTrue( WC()->cart->apply_coupon( $coupon->get_code() ) );
-
-		// Verifies 1 coupon has been applied to cart.
-		$this->assertEquals( 1, count( WC()->cart->get_applied_coupons() ) );
-	}
-
 
 	public function tearDown() {
 		// Reset the customer session data.
@@ -98,5 +105,4 @@ class Apply_Zipcode_Restriction_Test extends WP_UnitTestCase {
 		$this->customer->delete();
 		$this->coupon->delete();
 	}
-
 }
