+++ tests/phpunit/Integration/CheckoutExistingCustomerCouponTest.php
@@ -5,10 +5,10 @@ use WP_UnitTestCase;
 use WC_Helper_Customer;
 use WC_Helper_Coupon;
 use WC_Helper_Order;
-use WC_Coupon_Restrictions_Validation;
+use WC_Coupon_Restrictions_Validation_Checkout;
 
 class Checkout_Existing_Customer_Coupon_Test extends WP_UnitTestCase {
-
+	/** @var WC_Coupon */
 	public $coupon;
 
 	public function setUp() {
@@ -33,7 +33,7 @@ class Checkout_Existing_Customer_Coupon_Test extends WP_UnitTestCase {
 		);
 
 		// Run the post checkout validation.
-		$validation = new WC_Coupon_Restrictions_Validation();
+		$validation = new WC_Coupon_Restrictions_Validation_Checkout();
 		$validation->validate_coupons_after_checkout( $posted );
 
 		// Verifies 0 coupons are still in cart after checkout validation.
@@ -65,7 +65,7 @@ class Checkout_Existing_Customer_Coupon_Test extends WP_UnitTestCase {
 
 		// Run the post checkout validation.
 		// Coupon will be removed from cart because customer has previous purchases.
-		$validation = new WC_Coupon_Restrictions_Validation();
+		$validation = new WC_Coupon_Restrictions_Validation_Checkout();
 		$validation->validate_coupons_after_checkout( $posted );
 
 		// Verifies 1 coupon has been applied to cart.
@@ -83,5 +83,4 @@ class Checkout_Existing_Customer_Coupon_Test extends WP_UnitTestCase {
 		// Deletes the coupon.
 		$this->coupon->delete();
 	}
-
 }
