+++ tests/phpunit/Integration/CheckoutPostcodeRestrictionCouponTest.php
@@ -3,10 +3,14 @@ namespace WooCommerce_Coupon_Restrictions\Tests\Integration;
 
 use WP_UnitTestCase;
 use WC_Helper_Coupon;
+use WC_Coupon_Restrictions_Validation_Checkout;
 
-class Checkout_Zipcode_Restriction_Coupon_Test extends WP_UnitTestCase {
-
+class Checkout_Postcode_Restriction_Coupon_Test extends WP_UnitTestCase {
+	/** @var WC_Coupon */
 	public $coupon;
+
+	/** @var WC_Coupon_Restrictions_Validation_Checkout */
+	public $validation;
 
 	public function setUp() {
 		// Creates a coupon.
@@ -17,6 +21,9 @@ class Checkout_Zipcode_Restriction_Coupon_Test extends WP_UnitTestCase {
 		$coupon->save();
 
 		$this->coupon = $coupon;
+
+		// Inits the checkout validation class
+		$this->validation = new WC_Coupon_Restrictions_Validation_Checkout();
 	}
 
 	/**
@@ -37,7 +44,7 @@ class Checkout_Zipcode_Restriction_Coupon_Test extends WP_UnitTestCase {
 		WC()->cart->apply_coupon( $coupon->get_code() );
 
 		// Run the post checkout validation.
-		WC_Coupon_Restrictions()->validation->validate_coupons_after_checkout( $posted );
+		$this->validation->validate_coupons_after_checkout( $posted );
 
 		// Verifies 1 coupon is still in cart after checkout validation.
 		$this->assertEquals( 1, count( WC()->cart->get_applied_coupons() ) );
@@ -47,7 +54,7 @@ class Checkout_Zipcode_Restriction_Coupon_Test extends WP_UnitTestCase {
 		);
 
 		// Run the post checkout validation.
-		WC_Coupon_Restrictions()->validation->validate_coupons_after_checkout( $posted );
+		$this->validation->validate_coupons_after_checkout( $posted );
 
 		// Verifies 0 coupons in cart with invalid zipcode.
 		$this->assertEquals( 0, count( WC()->cart->get_applied_coupons() ) );
@@ -71,7 +78,7 @@ class Checkout_Zipcode_Restriction_Coupon_Test extends WP_UnitTestCase {
 		WC()->cart->apply_coupon( $coupon->get_code() );
 
 		// Run the post checkout validation.
-		WC_Coupon_Restrictions()->validation->validate_coupons_after_checkout( $posted );
+		$this->validation->validate_coupons_after_checkout( $posted );
 
 		// Verifies 1 coupon is still in cart after checkout validation.
 		$this->assertEquals( 1, count( WC()->cart->get_applied_coupons() ) );
@@ -95,7 +102,7 @@ class Checkout_Zipcode_Restriction_Coupon_Test extends WP_UnitTestCase {
 		WC()->cart->apply_coupon( $coupon->get_code() );
 
 		// Run the post checkout validation.
-		WC_Coupon_Restrictions()->validation->validate_coupons_after_checkout( $posted );
+		$this->validation->validate_coupons_after_checkout( $posted );
 
 		// Verifies 1 coupon is still in cart after checkout validation.
 		$this->assertEquals( 1, count( WC()->cart->get_applied_coupons() ) );
@@ -105,7 +112,7 @@ class Checkout_Zipcode_Restriction_Coupon_Test extends WP_UnitTestCase {
 		);
 
 		// Run the post checkout validation.
-		WC_Coupon_Restrictions()->validation->validate_coupons_after_checkout( $posted );
+		$this->validation->validate_coupons_after_checkout( $posted );
 
 		// Verifies 0 coupons in cart with invalid zipcode.
 		$this->assertEquals( 0, count( WC()->cart->get_applied_coupons() ) );
@@ -129,7 +136,7 @@ class Checkout_Zipcode_Restriction_Coupon_Test extends WP_UnitTestCase {
 		WC()->cart->apply_coupon( $coupon->get_code() );
 
 		// Run the post checkout validation.
-		WC_Coupon_Restrictions()->validation->validate_coupons_after_checkout( $posted );
+		$this->validation->validate_coupons_after_checkout( $posted );
 
 		// Verifies 1 coupon is still in cart after checkout validation.
 		$this->assertEquals( 1, count( WC()->cart->get_applied_coupons() ) );
@@ -139,7 +146,7 @@ class Checkout_Zipcode_Restriction_Coupon_Test extends WP_UnitTestCase {
 		);
 
 		// Run the post checkout validation.
-		WC_Coupon_Restrictions()->validation->validate_coupons_after_checkout( $posted );
+		$this->validation->validate_coupons_after_checkout( $posted );
 
 		// Verifies 0 coupons in cart with invalid zipcode.
 		$this->assertEquals( 0, count( WC()->cart->get_applied_coupons() ) );
