+++ tests/phpunit/Integration/CheckoutStateRestrictionCouponTest.php
@@ -3,9 +3,10 @@ namespace WooCommerce_Coupon_Restrictions\Tests\Integration;
 
 use WP_UnitTestCase;
 use WC_Helper_Coupon;
+use WC_Coupon_Restrictions_Validation_Checkout;
 
 class Checkout_State_Restriction_Coupon_Test extends WP_UnitTestCase {
-
+	/** @var WC_Coupon */
 	public $coupon;
 
 	public function setUp() {
@@ -39,7 +40,8 @@ class Checkout_State_Restriction_Coupon_Test extends WP_UnitTestCase {
 		WC()->cart->apply_coupon( $coupon->get_code() );
 
 		// Run the post checkout validation.
-		WC_Coupon_Restrictions()->validation->validate_coupons_after_checkout( $posted );
+		$validation = new WC_Coupon_Restrictions_Validation_Checkout();
+		$validation->validate_coupons_after_checkout( $posted );
 
 		// Verifies 1 coupon is still in cart after checkout validation.
 		$this->assertEquals( 1, count( WC()->cart->get_applied_coupons() ) );
@@ -67,7 +69,8 @@ class Checkout_State_Restriction_Coupon_Test extends WP_UnitTestCase {
 		WC()->cart->apply_coupon( $coupon->get_code() );
 
 		// Run the post checkout validation.
-		WC_Coupon_Restrictions()->validation->validate_coupons_after_checkout( $posted );
+		$validation = new WC_Coupon_Restrictions_Validation_Checkout();
+		$validation->validate_coupons_after_checkout( $posted );
 
 		// Verifies 0 coupons are still in cart after checkout validation.
 		$this->assertEquals( 0, count( WC()->cart->get_applied_coupons() ) );
@@ -81,5 +84,4 @@ class Checkout_State_Restriction_Coupon_Test extends WP_UnitTestCase {
 		// Deletes the coupon.
 		$this->coupon->delete();
 	}
-
 }
